pyproject.toml
[project]
name = "employeeanalyticshr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-ai-ml>=1.28.1",
    "azure-identity>=1.24.0",
    "faker>=37.5.3",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.5",
    "numpy>=2.3.2",
    "pandas>=2.3.2",
    "pandas-stubs==2.3.0.250703",
    "scikit-learn>=1.7.1",
    "seaborn>=0.13.2",
]
